+++ src/base/cs_mass_source_terms.h
@@ -48,7 +48,6 @@ BEGIN_C_DECLS
  * \param[in]     ncesmp        number of cells with mass source term
  * \param[in]     icetsm        source mass cells pointer (1-based numbering)
  * \param[in]     itpsmp        mass source type for the working variable
- *                              (see \ref cs_user_mass_source_terms)
  * \param[in]     volume        cells volume
  * \param[in]     pvara         variable value at time step beginning
  * \param[in]     smcelp        value of the variable associated with mass source