+++ src/cdo/cs_equation_param.c
@@ -3338,6 +3338,63 @@ cs_equation_add_volume_mass_injection_by_analytic(cs_equation_param_t   *eqp,
 
 /*----------------------------------------------------------------------------*/
 /*!
+ * \brief  Add a new volume mass injection definition source term by
+ *         initializing a cs_xdef_t structure, using a DoF function.
+ *
+ * \param[in, out] eqp       pointer to a cs_equation_param_t structure
+ * \param[in]      z_name    name of the associated zone (if NULL or "" if
+ *                           all cells are considered)
+ * \param[in]      loc_flag  where information is computed
+ * \param[in]      func      pointer to an analytical function
+ * \param[in]      input     NULL or pointer to a structure cast on-the-fly
+ *
+ * \return a pointer to the new \ref cs_xdef_t structure
+ */
+/*----------------------------------------------------------------------------*/
+
+cs_xdef_t *
+cs_equation_add_volume_mass_injection_by_dof_func(cs_equation_param_t  *eqp,
+                                                  const char           *z_name,
+                                                  cs_flag_t             loc_flag,
+                                                  cs_dof_func_t        *func,
+                                                  void                 *input)
+{
+  if (eqp == NULL)
+    bft_error(__FILE__, __LINE__, 0, "%s: %s\n", __func__, _err_empty_eqp);
+
+  /* Add a new cs_xdef_t structure */
+
+  int z_id = cs_get_vol_zone_id(z_name);
+
+  cs_flag_t  state_flag = 0, meta_flag = 0;
+  if (z_id == 0)
+    meta_flag |= CS_FLAG_FULL_LOC;
+
+  cs_xdef_dof_context_t  ac = {.func = func,
+                               .input = input,
+                               .free_input = NULL,
+                               .loc = loc_flag,
+                               .z_id = z_id};
+
+  cs_xdef_t  *d = cs_xdef_volume_create(CS_XDEF_BY_DOF_FUNCTION,
+                                        eqp->dim,
+                                        z_id,
+                                        state_flag,
+                                        meta_flag,
+                                        &ac);
+
+  int  new_id = eqp->n_volume_mass_injections;
+  eqp->n_volume_mass_injections += 1;
+  BFT_REALLOC(eqp->volume_mass_injections,
+              eqp->n_volume_mass_injections,
+              cs_xdef_t *);
+  eqp->volume_mass_injections[new_id] = d;
+
+  return d;
+}
+
+/*----------------------------------------------------------------------------*/
+/*!
  * \brief  Add an enforcement of the value of degrees of freedom located at
  *         the mesh vertices.
  *         The spatial discretization scheme for the given equation has to be